Update merged include paths for new code structure
