+++ src/runtime_src/core/pcie/tools/xbmgmt/cmd_reset.cpp
@@ -19,10 +19,10 @@
 #include <climits>
 #include <getopt.h>
 
-#include "scan.h"
-#include "xclbin.h"
 #include "xbmgmt.h"
-#include "mgmt-ioctl.h"
+#include "core/pcie/linux/scan.h"
+#include "core/include/xclbin.h"
+#include "core/pcie/driver/linux/include/mgmt-ioctl.h"
 
 const char *subCmdResetDesc = "Perform various flavors of reset on the device";
 const char *subCmdResetUsage = "--hot | --kernel | --ecc [--card bdf] [--force]";
